fix: warning from using LegacyVersion
